Do a config file based integration of evaldown.
Given the plugin's use of React and jsx syntax, a slightly more
involved inetgration of evaldown is needed using the transpilation
support to run the code through babel before it is evaluated. Use
a javascript module config file to set the transpileFn option - opt
to place the other options in there and pass --config to CLI tools.