Build only on publish rather than install to fix build.
This allows depending on the latest version of rollup that is not
compatible with node 8 but still build on that version (currently
still included in the build matrix). In order for build to succeed
in all the places it needs to (including locally) use node 12 for
the purposes of jobs in the build matrix and as the repo default.

This also reverts commit 6aac2983065783389f7536d26ae4f708a89193ca.